get project running in IntelliJ
I'm testing what new students will see when they clone the project for the first time. I want it to run as universally as possible, with as few steps as possible. To do this, I want to bundle as few of "Phil's personal settings" as I can to get the project running.