Add support for Visual Studio Code to project
I would like to have the option to view/edit/run the code in GitHub's online editor, and Visual Studio Code appears to be supported for this, at least somewhat. In addition, I would like to be able to consider teaching in Visual Studio Code. These settings files should allow the project to be opened and run out of VS Code for many users.